requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.3

# Google APIs
gspread==6.0.2
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# Web Scraping
selenium==4.17.2
webdriver-manager==4.0.1
beautifulsoup4==4.12.3
lxml==5.1.0

# HTTP & APIs
httpx==0.26.0
requests==2.31.0
aiohttp==3.9.3

# Google Trends
pytrends==4.9.2

# Data Processing
pandas==2.2.0
openpyxl==3.1.2

# Email
aiosmtplib==3.0.1

# Utilities
python-dotenv==1.0.1
pydantic>=2.10.0
pydantic-settings>=2.6.0
tenacity==8.2.3
itsdangerous>=2.0.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0
